Fixing liniting errors
