+++ lib/stacks/eks-blueprint-stack.ts
@@ -7,7 +7,7 @@ import { AutoScalingGroup } from '@aws-cdk/aws-autoscaling';
 import { Cluster, KubernetesVersion, Nodegroup } from '@aws-cdk/aws-eks';
 
 import { EC2ClusterProvider } from '../cluster-providers/ec2-cluster-provider';
-import { Team } from '../teams/team'
+import { Team } from '../teams'
 
 export class EksBlueprintProps {
